+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# REPLACE_GLOBALLY_WITH_NEXT_VERSION
+- PPI-624 - Improved error handling in after order process
+
 # 5.0.2
 - PPI-621 - Fix issue with the payment method overview missing in Shopware 6.4.7.0 or lower
 - PPI-623 - Fix an issue where a PayPal order cannot be created with discounts